requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.34.0
sqlalchemy==2.0.36
pdfplumber==0.11.4
paho-mqtt==2.1.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.20
